pyproject.toml
[tool.poetry]
name = "interfacetodns"
version = "1.0.0"
description = "Get interfaces from network devices and generate reverse DNS records."
authors = ["Rob Woodward <rob@emailplus.org>"]
license = "BSD-2-Clause"
readme = "README.md"
repository = "https://github.com/robwdwd/interfacetodns"
include = ["README.md", "LICENSE"]


[tool.poetry.dependencies]
python = "^3.7"
click = "^8.0.1"
easysnmp2 = "^0.3.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
flake8 = "^4.0.1"
pylint = "^2.9.5"
pydocstyle = "^6.1.1"
black = "^21.7b0"
pycodestyle = "^2.8.0"

[tool.poetry.scripts]
interfacetodns = "interfacetodns.cli:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
